+++ apps/avatax/src/modules/trpc/trpc-context.ts
@@ -1,11 +1,11 @@
 import { SALEOR_API_URL_HEADER, SALEOR_AUTHORIZATION_BEARER_HEADER } from "@saleor/app-sdk/headers";
 import { inferAsyncReturnType } from "@trpc/server";
-import * as trpcNext from "@trpc/server/adapters/next";
+import { FetchCreateContextFnOptions } from "@trpc/server/adapters/fetch";
 
-export const createTrpcContext = async ({ res, req }: trpcNext.CreateNextContextOptions) => {
+export const createTrpcContext = async ({ req }: FetchCreateContextFnOptions) => {
   return {
-    token: req.headers[SALEOR_AUTHORIZATION_BEARER_HEADER] as string | undefined,
-    saleorApiUrl: req.headers[SALEOR_API_URL_HEADER] as string | undefined,
+    token: req.headers.get(SALEOR_AUTHORIZATION_BEARER_HEADER) as string | undefined,
+    saleorApiUrl: req.headers.get(SALEOR_API_URL_HEADER) as string | undefined,
     appId: undefined as undefined | string,
   };
 };
